refactor(vcd): Restructuring Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
