+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/vcd/configuration/ConfigurationVcd.java
@@ -12,10 +12,12 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.configuration;
+package com.vmware.pscoe.iac.artifact.vcd.configuration;
 
 import java.util.Properties;
 
+import com.vmware.pscoe.iac.artifact.configuration.Configuration;
+import com.vmware.pscoe.iac.artifact.configuration.ConfigurationException;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
 
 public class ConfigurationVcd extends Configuration {
